// src/app/components/project/types.ts
export type ProjectObjectTypes = {
  projectid?: number;
  projectname?: string;
  projectdescription?: string;
  startdate?: string;
  enddate?: string;
  projectstatus?: string;
};

export type TaskObjectTypes = {
  taskid?: number;
  taskname?: string;
  taskdescription?: string;
  startdate?: string;
  enddate?: string;
};